Add log clearing

+++ views/app_logging.php
@@ -100,7 +100,7 @@ function render_logging_app_page($app,$urlm){
 
             } else {
 
-                if($currentProvision != null){
+                if($currentProvision != null && !isset($_POST['clearlogs'])){
 
                     SDC_LSLog::clearLogs($currentProvision->safeProvisionMeta('serviceid'));
                     $currentProvision->delete();
@@ -135,6 +135,10 @@ function render_logging_app_page($app,$urlm){
 
                     }
 
+                } else {
+                    if(isset($_POST['clearlogs'])){
+                        SDC_LSLog::clearLogs($currentProvision->safeProvisionMeta('serviceid'));
+                    }
                 }
 
             }
@@ -162,7 +166,7 @@ function render_logging_app_page($app,$urlm){
         <script type="text/javascript" src="<?php the_fileurl("static/js/jquery.min.js"); ?>"></script>
         <link rel="stylesheet" type="text/css" href="<?php the_fileurl("static/css/side-menu.css"); ?>">
         <link rel="stylesheet" type="text/css" href="<?php the_fileurl("static/css/top-menu.css"); ?>">
-        <link rel="stylesheet" type="text/css" href="<?php the_fileurl("static/css/app-logging-service-page.css"); ?>">
+        <link rel="stylesheet" type="text/css" href="<?php the_fileurl("static/css/app-logging-service-page.css?revision=3"); ?>">
         <script type="text/javascript" src="<?php the_fileurl("static/js/sidebar.js"); ?>"></script>
         <script type="text/javascript" src="<?php the_fileurl("static/js/greeting.js"); ?>"></script>
         <script type="text/javascript" src="<?php the_fileurl("static/js/app_logging_service_page.js"); ?>"></script>
@@ -232,6 +236,11 @@ function render_logging_app_page($app,$urlm){
             <div class="log-box">
                 <div class="lb-headline">
                     <p>Logs for this app</p>
+                    <form method="post">
+                        <?php $csrf->put(); ?>
+                        <input name="clearlogs" hidden>
+                        <button type="submit" class="clear-logs-btn"><svg style="height: 25px; width: 25px;"  viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-trash"><polyline points="3 6 5 6 21 6"></polyline><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path></svg></button>
+                    </form>
                 </div>
                 <div class="lb-box-iw">
                     <table id="log-table">
